[refactor] #73 TrackResponse 따로 파일로 분리
