+++ core/data/src/main/java/com/example/data/model/response/SearchTrackResponse.kt
@@ -1,6 +1,5 @@
 package com.example.data.model.response
 
-import com.example.domain.model.Track
 import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -19,25 +18,3 @@ data class SearchTrackResponse(
     @SerialName("items")
     val items: List<TrackResponse>
 )
-
-@Serializable
-data class TrackResponse(
-    @SerialName("trackId")
-    val trackId: String,
-    @SerialName("songTitle")
-    val songTitle: String,
-    @SerialName("artistName")
-    val artistName: String,
-    @SerialName("coverImg")
-    val coverImg: String,
-    @SerialName("isrc")
-    val isrc: String
-){
-    fun toDomain() = Track(
-        trackId = trackId,
-        songTitle = songTitle,
-        artistName = artistName,
-        coverImg = coverImg,
-        isrc = isrc
-    )
-}
